+++ runners/embedded/src/soc_nrf52840/types.rs
@@ -77,6 +77,7 @@ impl admin_app::Reboot for Reboot {
 	fn locked() -> bool { todo!() }
 }
 
+
 //////////////////////////////////////////////////////////////////////////////
 // Lower Interface (common definitions for individual boards)
